--- TODO.md
@@ -1,4 +1,0 @@
-# TODO
-
-1. Move frontend into components, only use install
-2. Refactor Backend
