+++ package.json
@@ -1,8 +1,8 @@
 {
     "name": "benweare.co.uk-server",
-    "version": "1.0.1",
+    "version": "1.1.0",
     "private": true,
-    "main": "./dist/index.js",
+    "main": "./build/index.js",
     "dependencies": {
         "@discordjs/voice": "^0.11.0",
         "axios": "^0.27.2",
@@ -20,9 +20,9 @@
     },
     "scripts": {
         "start": "node -r dotenv/config .",
-        "start-dev": "nodemon ./server/index.ts",
+        "start-dev": "nodemon ./src/index.ts",
         "build": "tsc -p ./ && tsc-alias -p ./tsconfig.json",
-        "test": "jest ./server/* --silent --runInBand"
+        "test": "jest ./src/* --silent --runInBand"
     },
     "eslintConfig": {
         "extends": [
