+++ src/handlers/contentHandler.ts
