+++ src/routes/weather/routes.ts
